chore(docs): refactor layout code
